refactor: rewrite update-appdata-xml to use xml2js and CommonJS
- Use xml2js for XML parsing/building instead of raw string manipulation,
  matching the pattern used by generateReleaseInfo.js and other scripts
- Convert from ESM (.mjs) to CommonJS (.js) to match all remaining scripts
- Extract duplicate regex-escape logic into shared escapeRegex helper
- Remove manual escapeXml function (xml2js handles escaping)
- Replace TOCTOU existsSync+readFileSync with try/catch on ENOENT
- Add proper error handling with top-level catch
